# src/openapi/openapi.gen.yaml
openapi: 3.0.0
info:
  title: veryExpress generated api server
  description: OpenApi Spec generated by veryExpress
  version: 1.0.0
paths:
  /api/customer:
    x-documentName: Customer
    post:
      operationId: postCustomer
      tags:
        - customer
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters: []
      requestBody:
        description: post Customer
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/postCustomerBody'
      responses:
        '201':
          description: Crerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/postCustomerResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
  /api/customer/search:
    x-documentName: Customer
    post:
      operationId: getListCustomer
      tags:
        - customer
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters: []
      requestBody:
        description: getList Customer
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getListCustomerBody'
      responses:
        '201':
          description: Crerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getListCustomerResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
  /api/customer/{id}:
    x-documentName: Customer
    get:
      operationId: getCustomer
      tags:
        - customer
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            x-format: ObjectId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getCustomerResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    patch:
      operationId: patchCustomer
      tags:
        - customer
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            x-format: ObjectId
      requestBody:
        description: patch Customer
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/patchCustomerBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/patchCustomerResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    delete:
      operationId: deleteCustomer
      tags:
        - customer
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            x-format: ObjectId
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AuthIndex:
      type: apiKey
      in: header
      name: X-Auth-Index
  schemas:
    getCustomerResponse:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        phone:
          type: string
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
    getListCustomerBody:
      type: object
      properties:
        _filter:
          type: object
        _select:
          type: array
          items:
            type: string
        _join:
          type: array
          items:
            type: string
      example:
        _filter:
          fieldName:
            $regex: str
            $options: i
        _select:
          - FieldName
        _join:
          - CollectionName
    getListCustomerResponse:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
          email:
            type: string
            format: email
          name:
            type: string
            maxLength: 64
          phone:
            type: string
          active:
            type: boolean
            default: true
          locale:
            type: string
            maxLength: 8
    postCustomerBody:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        phone:
          type: string
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
      required:
        - name
        - active
    postCustomerResponse:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        phone:
          type: string
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
    patchCustomerBody:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        phone:
          type: string
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
    patchCustomerResponse:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        phone:
          type: string
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
...